requirements.txt
biopython==1.75
cython  # install cython to install ruptures 
matplotlib==3.3.4
matplotlib-inline==0.1.6
numpy==1.19.5
pandas==2.2.0
ruptures==1.1.9
seaborn==0.11.2
